requirements.txt
customtkinter
speedtest-cli
pillow
